+++ lib/binary.dart
@@ -18,6 +18,8 @@
 /// [little endian]: https://en.wikipedia.org/wiki/Endianness
 library;
 
+export 'src/async.dart' show collectBytes;
+export 'src/bit_list.dart';
 export 'src/descriptor.dart';
 export 'src/extension.dart';
 export 'src/int16.dart';