+++ lib/src/uint16.dart
@@ -69,6 +69,9 @@ extension type const Uint16._(int _) implements FixedInt {
     max: 65535,
   );
 
+  /// Always `0`.
+  static const zero = Uint16.fromUnchecked(0);
+
   /// The minimum value that this type can represent.
   static const min = Uint16.fromUnchecked(0);
 
@@ -311,8 +314,26 @@ extension type const Uint16._(int _) implements FixedInt {
   /// ```
   Uint16 midpoint(Uint16 other) => Uint16.fromUnchecked(_.midpoint(other._));
 
-  /// Bits, from least significant to most significant.
-  Iterable<bool> get bits => _descriptor.bits(_);
+  /// Returns a list of bits representing this integer.
+  ///
+  /// The most significant bit is at index `0`.
+  ///
+  /// If [growable] is `true`, the returned list is a growable list.
+  ///
+  /// ## Performance
+  ///
+  /// A fixed-size list is a lightweight abstraction over a single integer, and
+  /// is can be inlined more easily by the compiler. Use a growable list only if
+  /// you need to modify the list.
+  ///
+  /// ## Example
+  ///
+  /// ```dart
+  /// Uint16(3).bits; // [true, true, false]
+  /// ```
+  List<bool> toBitList({bool growable = false}) {
+    return _descriptor.toBitList(_, growable: growable);
+  }
 
   /// Returns whether the n-th bit is set.
   bool nthBit(int n) => _.nthBit(n);